Fix Interactive window command handling
When I changed VsInteractiveWindow from is-a ToolWindowPane to has-a
ToolWindowPane, I broke its command handling.  I don't know why, since it
still implemented IOleCommandTarget and forwarded to the same method, but
revert that part of
https://github.com/dotnet/roslyn/commit/0e7a2b6e4fd3e79a78843319af965f5153bf21ad
until we figure out something better.

It's still not acceptable to upcast VsInteractiveWindow to ToolWindowPane,
so I'm leaving the IVsInteractiveWindow2 changes and we'll just have to
trust callers not to do it.
