+++ src/InteractiveWindow/VisualStudio/VisualStudioInteractiveWindow.csproj
@@ -129,7 +129,6 @@
     <Compile Include="CommandIds.cs" />
     <Compile Include="Guids.cs" />
     <Compile Include="IVsInteractiveWindow2.cs" />
-    <Compile Include="VsInteractiveWindowPane.cs" />
     <Compile Include="VsInteractiveWindowEditorFactoryService.cs" />
     <Compile Include="InteractiveWindowPackage.cs" />
     <Compile Include="IVsInteractiveWindow.cs" />
